requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
Authlib==1.5.1
bandit==1.8.3
boolean.py==4.0
boto3==1.37.4
botocore==1.37.4
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
cyclonedx-python-lib==8.9.0
defusedxml==0.7.1
dparse==0.6.4
exceptiongroup==1.2.2
fastapi==0.115.10
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
license-expression==30.4.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
nltk==3.9.1
packageurl-python==0.16.0
packaging==24.2
pbr==6.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.8.0
platformdirs==4.3.6
psutil==6.1.1
py-serializable==1.1.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pyhumps==3.8.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.3
safety==3.3.1
safety-schemas==0.0.11
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.46.0
stevedore==5.4.1
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
